+++ js/get-random-with-dot.js
@@ -1,0 +1,8 @@
+function getRandomWithDot(min, max, digits) {
+  if (min >= 0 && max >= 0 && max > min && min !== String && max !== String) {
+    return Number(((Math.random() * (max - min)) + min).toFixed(digits));
+  }
+  return 'Исходные данные неверны';
+}
+
+export {getRandomWithDot};
